Create Latest_File.vbs
The VBScript code iterates through files in a specified folder. Using a regular expression, it extracts dates in "YYYY-MM-DD" format from file names. The script then identifies the latest file by comparing these extracted dates. Finally, it displays the path of the latest file and the corresponding date. This code is useful for efficiently finding the most recent file based on date information in the file names.